Added function lsst::fw::kernel::convolveLinear, a special version of convolve that is faster for linear combination kernels. This is not yet well tested.
Changed a few Function methods to be non-virtual.

Minor change to how kernels compute kernel cofficients as a function of position.
This makes LinearCombinationKernel a bit more efficient by not allocating a needless
new kernel parameter array. This same speedup could be applied to AnalyticKernel
but only by allowing AnalyticKernel to get at the internal parameter vector of a Function
(or by changing Function to take params on the call line, which is dangerous unless length is checked).

Reluctantly made computeKernelParametersFromSpatialModel a public method of Kernel
so the function convolveLinear can get at it. This member function cannot affect
Kernel internals, but it can crash if not called correctly.