Add an API to BackgroundControl()/Background::getImage() that specifies the interpolation strategy in getImage
The old design mixed background estimation (in the Background ctor) and interpolation (also in the ctor),
and this is the first stage in disentangling them.  The old API is still supported (modulo the next
paragraph).

N.b. BackgroundControl can no longer be default-constructed. The 10x10 default wasn't generally useful, and
the new API (#2074) nolonger has a style argument that happens to disambiguate the string and enum versions.