Split Background into Background and BackgroundBase
The MaskedImage statsImage is really an implementation detail of this background model,
and not a property of the Background.  I accordingly added a base class and kept the
MaskedImage bits in the Background object.  Maybe I should have called them Background
and BackgroundMI or something.  Note that we now need to return a pointer from makeBackground,
hence the changes to the tests/examples