Fix bug in BilinearWarpingKernel
This fixes what I initially thought was a bug in afw::math::offsetImage(), but
turned out to be a bug in afw::math::BilinearWarpingKernel.

The bilinear warping code assumes that the x and y offsets of the interpolation
are between 0 and 1, and function evaluation is done at integer offsets 0,1.
This is the usual case, but there is also a code path from
afw::math::offsetImage() where the interpolation offset is between -1 and 0,
and function evaluation is done at integer offsets -1,0.

To reduce the potential for this to cause bugs in the future, I added error
checking by defining a virtual BilinearWarpingKernel::setKernelParameter()

This commit fixes the failing test case tests/offsetImage