Disable setting center pixel in cameraGeom.
The parts of cameraGeom that are known to work - getPixelFromPosition and getPositionFromPixel - do not make use of the _centerPixel data member; instead they compute the center pixel directly from the bounding boxes.  That makes being able to set the center pixel a potention source of errors, as it seems to affect some operations but does not in fact modify the all-important coordinate transformations.  This change removes setCenterPixel and the _centerPixel data member, and instead implements getCenterPixel by computing the center pixel of the bounding box, just as getPixelFromPosition and getPositionFromPixel do.

I have also removed methods that relied on the old, inconsistent definition of _centerPixel, including DetectorMosaic::findDetectorPixel, which purports to find the detector that contains a point in pixel coordinates.  At some level, that's clearly nonsensical, because e.g. all CCDs in a camera will have a (0,0) pixel.  I'm not sure it was used for anything other than circular unit tests.