Make it possible to round-trip a list of Backgrounds via disk
This involved adding a ctor to make a BackgroundMI from a statsImage
and the full-resolution image bounding box (n.b. the origin is
still ignored).   Because much task code still uses legacy API where
the interpolation style is specified in the ctor, not as arguments to
getImage where it's actually used, it seemed out of scope to expect
the list to be of tuples (Background, interpStyle, undersampleStyle)
so I use the values that were last used to get an Image.  This is
probably safe, but is a little ugly.