Remove dangerous implicit bool conversion from Color.
Equality comparison has also been added.

Overloading "operator bool()" is considered harmful, as this is an
implicit conversion that allows many other (usually meaningless)
operations on a class, by first converting it to bool.  In C++11, we
can use "explicit operator bool()", but don't think that's appropriate
for this class anyway.