Interpolate "correctly" in x even if we can't interpolate in y; #2297
We removed all NaNs when interpolating in y, but this fails if no good
values survive; we now set all values to NaN when this happens.  We then
go through another NaN-culling step when interpolating in x which means
we interpolate/extrapolate to fill out the image.

There's support in the code to set the value to a constant instead (see
#2825), but no API to use it.