Footprint: fix Footprint::transform (#2884)
The Footprint(Box2I const&, Box2I const&) constructor sets the Footprint
to cover the entire area of the first Box2I, but this is not what is
wanted, which is an empty Footprint.

This caused a large memory usage in forcedPhotCoadd.py because we were
adding spans to cover the entire area plus spans to cover the transformed
Footprint.