TanWcs: optimise distortPixel and undistortPixel (#2988)
TanWcs::undistortPixel figured in the profile for makeCoaddTempExp
(warping exposures to a skymap patch), contributing about 44 sec
of a total 155 sec.  The implementation was severely sub-optimal,
using pow() in tight loops to evaluate polynomials.  This optimised
implementation removes almost all of that 44 sec (to 0.01 sec).

The code affected is covered by tests/sipterms.cc, so it shouldn't
be necessary to introduce additional tests.