image,fits: consolidate FITS pickling code, and use for image classes (#3035)
Image pickling code didn't work on MaskedImage or Mask.  Stripped
out use of boost_picklable for image classes in favour of FITS-based
pickling (like Table; see #2344).  To do so, made the FITS-based
pickling code generally accessible; now anything can use it so long as
it can "writeFits" and "readFits".

Picklable image classes are: Image, Mask, MaskedImage, Exposure.

This adds new static "readFits" methods to the image classes that simply
forward to the proper Constructor methods.  While this increases the number
of interfaces, this parallels the operations in the Catalog classes (which
don't have proper Constructors that can read a FITS file).