Change to the _readTable routines in these 4 types plus BaseTable so that the AFW_TABLE_VERSION entry in the header is used to version each type of table. The getVersion() routine from BaseTable is used to get the version number during save, and it is written to the metadata of the table before the metadata is persisted.
On read, the metadata is searched at the beginning of FitsReader::_startRecords,
and the version number is set on the table then.  The table-getMetadata() is cleansed.
I also moved the removal code for AFW_TYPE to here.

Jim, please check the code to be sure that there are not unintended consquences from
moving the table->setMetadata to the middle of the _readTable routines