add NO_DATA mask plane, and use in warping/convolution
The EDGE mask plane was originally intended to mean "pixel near the
edge of the detector, so we haven't been able to properly convolve to
search for sources", but it was hijacked to mean "pixels in a warped
image that are off the edge of the original".  It has since had both
meanings, which should be distinct.

We introduce the NO_DATA mask plane to mean "pixels off the edge of
the original", and reserve "EDGE" to mean "pixels near the edge".
afw::math::edgePixel (used for pixels with no good inputs when
warping) now returns a value for pixels *off* the edge (should it
be renamed? but that's changing a public API).  Convolution now
masks pixels near the edge as EDGE if copying edge pixels, and
NO_DATA otherwise.