Replace misuse of getenv("AFWDATA_DIR") and fix outdated afwdata paths
afw was full of examples of std::string = getenv("AFWDATA_DIR")
followed by testing for a non-empty string. This doesn't work as expected
because getenv returns NULL if the environment variable does not exist.
Fixed by calling eups::productDir("afwdata") and explicitly checking
for the exception raised if the product is not setup.

Also, most of that code used outdated paths to files in AFWDATA.
In a few cases files were being read that no longer exist,
so minor additional changes were needed.