copyWithinFootprint: respect image size
copyWithinFootprint could produce a segfault (or at least corrupt
memory) when copying a Footprint that is larger than either of the
images.  Changed so that pixels outside either image are ignored
(i.e., it copies every pixel it is able to).  Added test of this
behaviour and documented the function's behaviour.

Conflicts:
	include/lsst/afw/detection/Footprint.h

This is a cherry-pick from HSC-1004's commit:
https://github.com/HyperSuprime-Cam/afw/commit/76f3706f6688b23d5b0c71e66af3e94095a9f821