Add variable-length array fields to afw::table
We use Array<T> with a size of zero to indicate variable-length fields.
It'd probably be better to use a completely different tag type, but
Boost.Variant puts an upper limit on how many tags we have, so until
we rewrite afw::table to avoid it, we need to limit how many we use.

Conflicts:
	include/lsst/afw/table/FieldBase.h
	src/table/Schema.cc
	tests/testSimpleTable.py