More careful handling of mismatched arrays in footprint-merge tests
Some of this test code wasn't quite testing what it claimed to be testing,
because the arrays being passed in were not all the same size, and zip()
silently iterates over the smallest of its arguments.  The removal of
some of the test utility code in this change is just to make it so we have
to worry about that problem in fewer places.

Also changed the failure mode to an exception instead of a warning
and no-op; seems like a logic error that should be fatal in order to
better catch bugs in higher-level code.