Functions and methods that create Footprints must preserve Peak schemas
Many of the operations that created new Footprints from old ones
internally used the Footprint default constructor, not the copy
constructor, and that meant they weren't transferring the new
peaks properly.

Conflicts:
	src/detection/Footprint.cc

Note that recent improvements on the Footprint grow and shrink
algorithms require a separate commit to ensure the peaks are
being transferred and updated properly.