Cleanup all relevant aliases when defining slots.
We now erase all matching aliases when defining a slot, instead of
assuming that the slot is defined by only one alias.  This does cause
a bit of unnecessary processing (as the SourceTable has to handle
callbacks for both erasing and setting aliases), but we don't expect
this code to ever be called in a context where that matters.