Add Functor and SeparableXYTransform classes to support modeling of sensor effects.
Functor is an abstract class that SeparableXYTransform takes as
arguments in its constructor.  Distinct Functor subclass objects are
used to model the transformations in the x- and y-directions,
respectively.  Functor provides an inverse(...) member function that
calls the subclass' concrete implementation of the Functor::derivative(...)
pure virtual function to compute the inverse of the 1D function using
Newton-Raphson.  LinearFunctor is provided as an example subclass.