Change Background to use the xy0 of the image bbox
The Background and BackgroundMI objects ignored the xy0 of the image
from which they were constructed. This code modifies these classes
to record the xy0, so that the background images constructed from an image
have the same xy0 as the image.

It appears that ignoring the image's xy0 was an accident based
on a misunderstanding about the old default behavior of Image.getBBox(),
which used to return a LOCAL bounding box (xy0=0,0). Much of the existing
code seems to be based on the bounding box possibly having a non-zero xy0.
Nonetheless, simply changing the way the bounding box was read from the image
was insufficient and I had to make additional changes. There were two possible
approaches:
- Use PARENT coordinates everywhere. This requires changingmore code, but results in all
variables using the same coordinate system (PARENT coordinates), so I felt it was
worth trying. Unfortunately I was not able to get it to work.
- Continue to use LOCAL coordinates for all the math and just apply the XY0
after computing the background image. This is a minimally invasive approach,
but result in the code using two coordinate systems, which is potentially confusing.
Nonetheless this is the approach I took.

tests/background.py includes two new tests for these changes. I also made
other changes to this test, including:
- Minor changes to make the file pass the pyflakes linter with no warnings
- Use @skipIt for tests if afwdata is not setup, as requested by the reviewer

Improve the background.py unit test as per John S's review