Add temporary debugging output for makeStatistics
The debugging output shows which vector-like makeStatistics variant
is being called, and it shows that std::vector is preferred to
ndarray for Python when both apply. We want ndarray to be preferred,
as it does not copy data, but so far I have not figured out how to do that.
(Swapping the declaration order to list the ndarray versions first did not help).
Also the unit16/64 makeStatisitcs are not found when using numpy.uint16/64 dtypes,
so the SWIG wrapper still needs some work.