Add hasTransform and getTransform to cameraGeom Detector
Add overloaded member functions hasTransform and getTransform to cameraGeom Detector,
which accept a CameraSys or a CameraSysPrefix. This greatly simplifies getting
appropriate transforms from the detector's transform map, since one need not explicitly
first generate a CameraSys for that detector.
Added a test for the new functions to testDetector.py.
Made a minor improvement to documentation in TransformMap.h

Fix grammatical errors in two doc strings.