Refactor FITS binary table reading.
This is a major refactoring aimed at cleaning up some code that had
gotten messy after a lot of extensions the original design had never
been expected to support.  The final straw was breaking the assumption
that a Schema object provides provides enough per-column information
to read records from disk, which is necessary to support backwards
compatible reading of version=0 Schemas after we remove support for
version 0 objects in memory.

This change should be almost entirely invisible to the user, with the
only exception being the removal of the (unused) Schema constructor
from PropertyList.  Now that we've recognized that Table objects need
to process the metadata as well (and that the Schema that's loaded
from the metadata depends on the Table subclass), supporting
construction of Schemas from metadata isolation no longer makes sense.