Avoid premature removal of AFW_TYPE FITS keyword
This keyword is needed when the Schema is created to determine
the table version - if it's present, but no version number is, we
should use version 0, as that means it was written before we
started writing version numbers.  In contrast, if it's not present,
we assume afw didn't create it, so we should use the current
version.