Implemented Linear and ZScale transformations, HSC-1206
Linear was already available with Q=0, but this way is clearer; the
default is to use min/max limits.

I also added AsinhZScaleMapping which uses the zscale code to set the
stretch.

I did not port the log mappings, as the C++ is not really something
I'd like to see in the long-term stack (e.g. reinterpret_casts between
ints and doubles). It would be easy to add a LogMapper in pure python,
and if this is needed please open a new issue.

The ZScale code comes from Mineo-san's fitsthumb, but doesn't give quite
the same upper limit as ds9.  I'm not sure where the discrepancy is, but
I don't think it matters.