Adapt test code to avoid import of meas_algorithms
The tests/footprintMergeCatalog.py code made use of meas_algorithms
DoubleGaussianPsf to create a psf for a generated image.  Since afw
does not require meas_algorithms as a build dependency, it cannot be
assumed that the latter is set up and available.  The psf is now
created using afw.detection's GaussianPsf.