Change tests/coord.py to use print as a function
Updated coord.py to use from __future__ import print_function
and updated print statements accordingly.
Also there were a number of cases where "s" was being set to a string,
which was then printed; I modified the code to just print the string
without setting it to a temporary variable.