Gracefully fail when a persisted object fails to load more than once
Prior to this fix, the first failed load would throw an exception
but leave the cache of loaded objects in a state that later loads
would assume meant the object had loaded successfully, leading to
an assertion failure.  Now, later loads will detect this cache
state and just throw again.