Fix performance regression in Footprint dilation.
The earlier RLE-based Footprint dilation routine relied on
Footprint::normalize() to weed out duplicate and overlapping Spans. Although
good performance was obtained for small Footprints and structuring elements,
this could result in a significant performance regression when a large number
of Spans were created during dilation.

Here, the algorithm is adjusted so that the list of (proto-)Spans is
maintained in a normalized state as the Footprint is grown, thereby avoiding
the bottleneck.

Conflicts:
	src/detection/Footprint.cc: from C++11 removal
	tests/footprint1.py: from additional test