Footprint: add method findEdgePixels
This is an operation used in the deblender, optimized (not calling
Footprint::contains twice per pixel) and moved here in case it may be of
general use.  Edge pixels are identified as pixels that are on the edge
horizontally or have nothing above or below them.  We keep a representation
of the row before and after to save on calling Footprint::contains for
each pixel.

Conflicts:
	include/lsst/afw/detection/Footprint.h
	tests/footprint1.py

This is a cherry-pick from HSC-1274's commit:
https://github.com/HyperSuprime-Cam/afw/commit/1c6f30decf5371dd855f1d0225a01e64ec2ab143