interpolate: fix off-by-one bug in InterpolateConstant
Nate Lust found an "invalid read" in the "recenter" function
when running Valgrind on a test of PSFEx.  Inspection of the
code found that we were running off the end of the array.
This error is not (in general) reproduced by running valgrind
on afw's interpolate.py test, probably because we're getting
more space back from the standard library than we ask for.
Changing the test to use 8 or 16 values allows valgrind to
see the error, which goes away with this fix, but I've not
modified the test.