Statistics: assert that images have the same size
makeStatistics() assumes that image[0,0] and mask[0,0] are the same position.
If mask is bigger than image, but all the unmasked area falls outside the
image given this alignment, then attempting to calculate statistics will
produce a NaN. We avoid this possibility by refusing to proceed if the image
and mask do not have the same dimensions.