Improve the way formatters log messages
Formatters were logging using a name, but this is slower
than using a Log object. So I added a Log object with a suitable name
to each class and use that for logging.
Note: in choosing the name I adopted the namespace of the object
being persisted (e.g. afw.image.ImageFormatter instead of
afw.ImageFormatter), instead the location of the formatter itself,
since the latter seemed more arbitrary and less predictable.