Initialized Sphinx package docs
Product of sphinx-quickstart; with customizations.

Customized conf.py:

- cleanup and remove settings for unsupported build targets
- breathe config via documenteer
- use sphinx_rtd_theme

Eventually this configuration should be centralized into documenteer so
that it can be applied and maintained uniformly across all packages
(with YAML as a customization tool).

Also trimmed Sphinx Makefile to supported targets.

Add local .gitignore for Sphinx build directory

For DM-4195.