Use sconsUtils to locate package directory rather than lsst.utils
lsst.utils can not be loaded by scons on El Capitan because
DYLD_LIBRARY_PATH is not reliable. sconsUtils already knows
how to locate packages so remove the dependency on lsst.utils
and use the sconsUtils implementation.

Also remove the warning concerning daf_persistence. This warning
does not prevent the tests themselves warning or disable any
tests and will not work on El Capitan.

The afw tests changes are a temporary fix until the tests themselves
can be fixed to check for afwdata. Disabling the tests at the scons
level prevents metrics tools from understanding exactly which tests
are being skipped. All test skipping based on environment should
be within the test logic.