Speed up warping if src and dest wcs have same csys and equinox
Added an optimization to XYTransformFromWcsPair's transforms
for the situation that the source and destination WCS
have the same coordinate system and date of equinox.
This speeds up warping by roughly a factor of two
(when not using interpolation) when applicable.

I also added three methods to Wcs: getCoordSystem, getEquinox
and isSameSkySystem and a unit test for these.

Note that the original warping code (before this ticket)
mis-handled the situation where source and destination
coordinate system or equinox did not match.
I still have to add a unit test for that.