Allow for case where astronomy_net_data not setup
The matchesToCatalog() function attemps to set a "REFCAT" key in the
match metadata by looking up the basename of the astronomy_net_data
package directory.  This would throw in the case where
astronomy_net_data is not setup, so wrap it in a try statement and
set "REFCAT" to "NOT_SET" in the case where it is not setup.