Fix bug in transforming WCS
There was a bug in the code that transforms WCS objects for assembly.
Specifically, it shifted the reference pixel all the way to the origin
when accounting for trimming when, in reality, one wants to just shift
the reference pixel the difference between the amp origin in the
untrimmed coordinates relativel to the trimmed coordinates.