tests: fix matching test that failed on OSX
The catalogs being matched had multiple exact matches, and so which
particular match was output appears to depend on the compiler; but
assumptions on which particular match should be output had been made
in the test, which are apparently violated on OSX with clang.

Removed the multiple exact matches and clarified the tests, including
the addition of comments explaining how the tests are set up and what
the results of particular tests should be.