Add method to get writeable bool column arrays.
We normally return read-only bool arrays when accessing
Flag columns in Python, even though the arrays are copies
and safe to edit, to prevent users from thinking they can
modify Flag field values via the array interface.

However, in an astropy.table view, users will generally
asking for a copy, and in that case we don't want to
restrict what they can do with it.