Implement declared but undefined method in Ellipse::GridTransform.
The fact that this lacked an implementation has somehow escaped notice
for years.  That's's probably related to the fact that it isn't
accessible from Python, where the details of the GridTransform don't
quite translate and hence we go through a less efficient but otherwise
equivalent code path.  The new unit test compares those two code paths.