Change AmpInfoRecord saturation level to double
It turns out that IsrTask masks saturated pixels after the image
is converted to floating point. Furthermore, the task does no masking
if the saturation level is `nan`. However, the saturation level
was an integer, so I changed the saturation level to a float
to match the task code.

Also document `nan` to mean no level in the help for the field.