Import fixes on python3
If lsst.display is not loadable you get a SystemError on Python3
when trying to do an import relative to it. You do not get
ImportError. Trap for both errors.

Also check that _impl attribute exists before using it. Sometimes
the close() method can be called without the object completing
initialization.