Skip test if afwdata is not available
There is logic in the SConscript file to disable tests that use
afwdata but this logic does not get enabld when testExectuables.py
is run or if a user runs the test from the command line. This
patch adds additional skipping logic inside the test without
affecting the scons logic.