More careful screening of arguments to selfInterpolate()
This is the minimal change needed to fix the intermittent
failure that spawned this ticket.  selfInterpolate() was
not noticing if users passed in a negative index (which
makes no sense) or asked for nData nearest neighbors (not
possible with selfInterpolate, since selfInterpolate ignores
the point at which interpolation is requested when assembling
the list of neighbors (so nData-1 is the largest number of
neighbors you can request).