Introduce SpanSet Class
SpanSet is a class which holds a collection of spans to represent a collection
of spatial indexes. This class has associated methods for working with SpanSets
such as intersection, logical difference, and union. It also contains
functionality which takes in a functor, and containers to operate on, and
applies the functor to the containers at each of the points defined in the span
set.