Make computeImageStack work
The computeImageStack function in stack.cc was not able to stack objs
of type afwImage.Image<PixelT>. The problem seemed to stem from the use
of a MaskedVector to hold the pixels being stacked. this commit
switches the computeImageStack to using a simple std::vector instead
which fixes the bug.