Fix pybind11 downcasting by avoiding invisible subclasses.
Table-Record pairs had previously been implemented as trivial
subclasses visible only to source files so we could use make_shared,
but this breaks pybind11's clever ability to downcast base class
pointers to the most derived type.  We can get that working again
by dropping the trivial subclasses and just using 'new'.

This also adds 'override' to a few other methods, because using it on
one method in a class and not others leads to warnings with clang.