wcs: reduce pickle size
The pickle size was increased in commit 74b8006.  The Exposure includes
multiple copies of the Wcs (in the header of each of image, mask,
variance, plus the archive version). But we don't have to embed the
Wcs in an Exposure --- we can just write the Wcs to FITS directly.
This decreases the size of the Wcs pickle by a factor of 5.

Rebased from LSST: pybind changes converted to SWIG changes.