Change 2 Utils funcs to take &, not shared_ptr
Many functions in lsst/afw/formatters/Utils.h take
a shared pointer to a property set when a reference will do.
I changed two that I would like to call from the new WCS code
without creating a shared pointer to a PropertySet.
These two functions are trivial to change because they
are called in only a very few places, all in afw.
Many other functions could be changed in the same way,
but at some point one runs into problems
constructing a `LogicalLocation` so I kept it to a minimum.