Enable default copy; disable move assigment
Transform and SkyWcs had the move assignment operator enabled
(default implementation) but I want them to be immutable
so I'm disabling that.
They also had the copy operator disabled, but I'm enabling that
(the default implementation) since they are safe and cheap to copy.
So now copy and move are enabled but assignment is disabled.