Fix limit handling in Angle.wrapNear
A test for the result of Angle.wrapNear being too small
applied the incorrect correction when the wrapped result was positive.
Added a test that shows the problem (it was not easy to trigger,
but setting wrapRef to 5 did the trick).
Adding wrapRef caused indentation to be so deep that I switched
to using itertools.product to tame it.