Fix signed vs. unsigned warnings.
Most of the warnings in GaussianProcess.cc were patched with
on-the-spot casts. While changing the internal members to unsigned
makes more sense a priori, the current code uses -1 as a null value,
and it is very difficult to maintain KdTree's class invariants during
the conversion. It may be less work to just replace the implementation
of KdTree with an object-oriented rather than array-oriented
implementation.