Change mask type to int32
Changed Mask pixel type(def) to int32, adding alias to MaskPixel
in python as well. This change triggered a renaming of the python
template alias from MaskU to MaskX to avoid type confusion.
Because of the new pybind11 continue class abstract base classes
this new name could be mostly hidden behind the abc name "Mask",
as MaskX is now the default type to construct when calling the
abc.