Improve CopyValue.operator() for var-length arr
CopyValue.operator() for a variable-length array presently
only checks that output is variable length if input is variable length;
I changed it to always check that both match in this way.
Also if they don't match I changed an assert to an exception,
since it is easy to provide non-matching schema and one can
do so without corrupting memory.
Question: should we check that the sizes match? That would be
more thorough, but perhaps it is more than we need?