Clean up BoundedField pybind11 wrapper
and update the ChebyshevBoundedField wrapper accordingly.

Wrap `__ne__` for `BoundedField` to eliminate the need to wrap
it in subclasses. This was especially nasty because the
C++ method `operator!=` was only defined in `BoundedField`,
so it was easy to miss in wrappers of all subclasses,
and failure to wrap it resulted in errors that only showed
up in Python 2, not Python 3.
Wrap `__eq__` for `BoundedField`, even though it has no
implementation. This is best practice and may eliminate the need
to wrap it in subclasses, but I'm not going to test that.
Wrap `__rmul__` for BoundedField to eliminate the need to wrap
it in subclasses.

Update the `ChebyshevBoundedField` pybind11 wrapper accordingly:
remove the wrapper for `__ne__` and `__rmul__`.