Stop storing xy0 in SkyWcs
It turns out to be impractical to have SkyWcs store xy0,
since when one shifts the WCS it's not clear if it shifts
the parent or local frame. Better to emulate the Wcs classes
and have external code shift the SkyWcs as required.

In a related vein, stop AST from stripping keywords for WCS "A"
when it reads FITS metadata, since LSST uses some of those keywords
to store XY0 and SkyWcs has no use for the information.