fits: fix doxygen warnings
fitsCompression.cc is causing doxygen warnings:

    documented symbol `ImageScale lsst::afw::fits::ImageScalingOptions::determine' was not declared or defined.

This appears to be due to a bug in doxygen (perhaps having to do with
underscores at the end of variable names?). To deal with this, corrected
the grouping markup for ImageScalingOptions::determine and put
fitsCompression.cc on the doxygen exclusion list. This doesn't produce
correct docs (e.g., ImageScalingOptions::determine isn't listed as a
member function) but it's no worse than it was before, more correct and
doesn't produce a warning.