stack: set mask bit when pixel is clipped or masked
We want to know when a pixel is clipped or masked (because on coadds that
means the CoaddPsf won't be accurate), so have a bit set when this occurs.
Uses the new NCLIPPED and NMASKED statistics.

Some mask bits may be excused from triggering this, so as not to mask
as clipped/masked pixels that shouldn't ever be considered for inclusion
in the stack (e.g., NO_DATA; as opposed to pixels that are not being
included because the are in the imaged area but corrupted somehow).