Allow SpherePoint.bearingTo and offset at poles
For a SpherePoint constructed from a 3-vector, set longitude to 0
if at a pole and document the same. This allows bearingTo and offset
to be well defined for all cases.

Also change operator== to compare longitude even at the poles,
since different longitudes result in different behavior for bearingTo
and offset even at the poles.