fits: extend list of header keywords to ignore
There are some keywords that, if we write, it will corrupt the
FITS file because cfitsio takes care of them; these typically
have to do with core FITS concepts (size, column definitions,
compression) that we handle explicitly through cfitsio
interfaces. And the user shouldn't care about them either.
However, there are some keywords that need to leak through
our keyword stripping because afw::table::io needs to see
them in order to build a Schema; we just won't allow those
to be written.

This includes refactoring the 'isKeyIgnored' function to make
the list of keywords more easily extensible.