Fix error in test_skyWcs.py
The test of SkyWcs.getTanWcs was only passing because TAN-SIP
was not being written out by AST, a problem fixed on this ticket.
For now remove the test for SkyWcs.getTanWcs();
tickets/DM-10765 has a much better test, but it relies on
a function that that is not available on master.
The new test is checkNonFitsWcs and it is called by
testMakeModifiedWcsNoActualPixels and
testMakeModifiedWcsWithActualPixels