move test camera data to cameraGeom/testData/
When building Docker images, SQuARE policy to to delete
the tests/ directory of a package after it is built.
Unfortunately, the data for test camera existed in tests/,
even though the code to create the test camera existed in
python/.  Several sims packages rely on being able to
instantiate test camera to run their unit tests.  When
building these packages on top of a Docker image of a DM
weekly, the tests would fail because the data had been
deleted.  By moving the data into python/, we make sure
that it will be preserved for use by downstream packages.