Transform: calculate Mapping at construction time
Much of the runtime performance degradation from SkyWcs appears to be
due to AST rebuilding the transformation every time we want to use it.
By calculating and caching the '_frameSet->getMapping' at construction
time, and using that for the coordinate transformations, we can cut
the corner and save a LOT of time.

Thanks to Russell Owen for the idea.