Make Transform contain only an ast::Mapping
instead of an ast::Mapping and an ast::FrameSet, as per RFC-469.
Rename Transform.getFrameSet() to getMapping().
Add an internal ast::FrameDict to SkyWcs, since it was using Transform
to store its FrameDict. There are no API changes to SkyWcs.
Update and clarify documentation.
Update other code and unit tests as needed.