Make warpImage call imagesOverlap
Make warpImage call imagesOverlap instead of isSameObject
(which didn't compare the right thing and didn't even try
to detect overlap).

Document that the source and destination must not overlap
(formerly we only attempted to make sure they were not the same image).

Fix the unit test that checked that overlapping images could not
be warped onto each other. It was completely broken.