Fix an error in footPrintSet pybind11
Automatic conversion of ``std::shared_ptr<T>`` is not possible when ``T`` is
not directly registered with ``py::class_<T>`` (e.g. ``std::shared_ptr<int>``
or ``std::shared_ptr<std::vector<T>>`` are not automatically convertible).

In this case all we need to do is get rid of an overload that was
not being used. But I also made the existing overload more efficient
by using std::move (as suggested by Pim Schellart) and simplified
the associated code comments.