Rename suffix Sigma to Err in CovarianceMatrixKey
in order to comply with RFC-333 (note that we do not have a
use case for writing covariance matrices with Sigma, just Err).
Increment the schema version number from 1 to 2.
Update the code that converts version 0 schemas.
Add code to read version 1 schemas.
Change variable names and typedefs accordingly.
Update the help for CovarianceMatrixKey.
Update the unit tests as required.