fix table.Schema.__contains__
Following the pybind11 upgrade, error_already_set::clear doesn't do
anything (it's deprecated), so it can lead to a python GIL deadlock
when the object is destroyed. Instead, hand the error back to python
and clear it (so we're not holding the error and no error is set);
this avoids the deadlock.