Replace Source.h.m4 with the generated Source.h
Remove SConscript used by m4

After discussion with Jim Bosch, we agreed it was probably easier to use
modern smart editors to make bulk changes to Source.h instead of having
someone in the project try to remember the m4 syntax. Also, we hope to make
that header much simpler by removing a bunch of the deprecated code.