Only write BLANK for quantized floating-point images.
BLANK is used to represent missing values in integer images only
(NaN is always used for floating point).  But our in-memory integer
images can never have missing values, so the only place we need
BLANK is when quantizing (for compression) a floating-point image
into an integer image.

Because ZBLANK is what BLANK should be after decompression, we should
never write it.