Add test for flux field units issue of DM-16068
The _flux->_instFlux alias setting for backwards compatibility with
catalogs with version < 3 relies on flux fields having their units
set properly.  Prior to a resolution of DM-16068, the units were not
getting set for several CModel flux fields and one deblender field
(deblend_psfFlux), and thus were not getting the _flux->_instFlux
aliases set.  In turn, the very reading of a of a coadd catalog
fails as the CModel flux is what slot_ModelFlux points to, and
slot fields are always searched for.

NOTE: this test will (is meant to) fail on the read until DM-16068
      is resolved.