Add hash for equality-comparable types in afw::coord and afw::geom.
The affected types are Observatory, Polygon, Span, and Weather.

Types with complex inheritance relationships have been excluded
because of the difficulty of using polymorphism with the std::hash
template. Types that are noncopyable have been excluded because almost
any situation where hashes are useful requires copyable types, and
because handling noncopyable test objects is *extremely* difficult.