Stop trying to remember raw pointers in OutputArchive.
The previous behavior attempted to remember all pointers saved and
return the same archive ID when the same pointer was saved twice; this
permitted complex data structures that had multiple pointers to the
same object to be reconstructed exactly.

Unfortunately, if temporary objects are created, saved, and then
destroyed during the lifetime of a single archive, the pointer
addresses could be reused, fooling the archive into thinking a new,
different object had already been saved.

To avoid this, we now only attempt to remember shared_ptrs, and use a
weak_ptr map in the archive to ensure new objects are never mistaken
for dead ones (without letting the archive actually keep objects
alive).  That means we can only round-trip pointer relationships when
they're shared_ptr relationships, but happily that's all we have right
now (outside of test code).