Use NumPy stringification to generate test results.
Previously, we were checking against hard-coded “guesses” of what NumPy might
return for a given array. That conflates two different tests:

- Are we correctly delegating formatting to NumPy?
- Is NumPy actually giving us what we want?

This modification means we are only testing the former, which, I suggest, is
appropriate. We could consider adding a test to verify that some future
version of NumPy doesn't change its output to something which is unhelpful to
us, but I suggest that that is out of scope for this work.