Fix removeMaskPlane function and add unittest
The removeMaskPlane was broken in two ways:

1) the removeMaskPlaneDict() function called is not an attribute of the
mask being updated, so any call to MultibandMask's removeMaskPlane()
results in an AttributeError.  The fix is to update the function name
from removeMaskPlaneDict() to removeMaskPlane().

2) the function ignores the variable "name", i.e. does not pass it to the
call to [sic!] removeMaskPlaneDict().  Having corrected the function name
to removeMaskPlane(), we then get a TypeError.  The fix here is to pass
name into removeMaskPlane(name).

These issues were not caught because removeMaskPlane() was not being
exercised in any unittest, so one is added here.