Add makeTestExposure.py and an unversioned Exposure
We didn't have a test Exposure file with enough fields filled in to be useful,
so I wrote this to make it easier to generate such files in the future,
particularly as we start changing the contents/version of Exposure.

The file is an "unversioned" Exposure, created before there was any
information about file version in the Exposure output.