Add workaround for PropertyX lack of support for undefined arrays
PropertySet (and List) can handle arrays of undefined values
but can not handle the addition of an undefined value to an
existing array, or the addition of a good value to an
undefined array.

With this change we make the following allowances when reading
FITS headers that may have multiple keywords, some of which
may be undef:

* If a defined value is being added to a key that exists but
  is undefined, the defined value will override.
* If an undef value is being added to a key that exists but
  is has one or more defined values, the undef value will be
  dropped.

A test has been added to ensure that this is happening.